Remove onBack methods
