+++ lib/features/add_book/widgets/add_book_submit_button.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 import '../../common_components/common_form_submit_button.dart';
 import '../bloc/add_book_form_bloc.dart';
@@ -13,7 +14,17 @@ class AddBookSubmitButton extends StatelessWidget {
       alignment: Alignment.centerRight,
       child: CommonFormSubmitButton(
         onPressed: BlocProvider.of<AddBookFormCubit>(context).submit,
-        onSuccess: () => Navigator.of(context).pop(true),
+        onSuccess: () {
+          Navigator.of(context).pop();
+          ScaffoldMessenger.of(context).showSnackBar(SnackBar(
+            content: Text(AppLocalizations.of(context)!.addBookSuccessfully),
+          ));
+        },
+        onFailed: (_) {
+          ScaffoldMessenger.of(context).showSnackBar(SnackBar(
+            content: Text(AppLocalizations.of(context)!.addBookFailed),
+          ));
+        },
       ),
     );
   }
