+++ lib/features/table_of_contents/widgets/toc_edit_book_button.dart
@@ -17,7 +17,7 @@ class TocEditBookButton extends StatelessWidget {
     final BookData bookData = BookData.fromName(bookName);
     return IconButton(
       onPressed: () {
-        _navigateToEditBook(context, bookData).then((newData) => _onPopBack(context, newData));
+        _navigateToEditBook(context, bookData).then((_) => BlocProvider.of<TocCubit>(context).refresh());
       },
       icon: Icon(
         Icons.edit_rounded,
@@ -50,23 +50,6 @@ class TocEditBookButton extends StatelessWidget {
             );
           },
         );
-    }
-  }
-
-  /// Handle the result of editing the book
-  void _onPopBack(BuildContext context, dynamic newData) {
-    if (newData is BookData) {
-      /// Refresh the table of contents
-      BlocProvider.of<TocCubit>(context).setDirty();
-      BlocProvider.of<TocCubit>(context).refresh(newData: newData);
-      ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-        content: Text(AppLocalizations.of(context)!.editBookSuccessfully),
-      ));
-    } else if (newData == false) {
-      /// Failed to edit the book
-      ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-        content: Text(AppLocalizations.of(context)!.editBookFailed),
-      ));
     }
   }
 }
